requirements.txt
patchright==1.52.4
PyYAML==6.0.2